requirements.txt
pandas==2.3.0
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
reportlab==4.4.1
python-dotenv==1.1.0
fastapi==0.116.1
